pyproject.toml
[project]
name = "campus-rush"
version = "0.1.0"
description = "赶早八 · Campus Rush - A university campus parkour game"
requires-python = ">=3.12"
dependencies = ["pygame-ce>=2.5.0"]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
